more refactor for cleaner flow. also more REST stuff, and more OpenDips stuff

+++ XcaXds.WebService/Controllers/RestfulRegistryRepositoryController.cs
@@ -1,4 +1,8 @@
 ﻿using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Logging;
+using Microsoft.FeatureManagement;
+using System.Diagnostics;
+using System.Reflection.Metadata;
 using XcaXds.Commons.Models.Custom.DocumentEntry;
 using XcaXds.Source.Services;
 using XcaXds.WebService.Attributes;
@@ -16,23 +20,36 @@ public class RestfulRegistryRepositoryController : ControllerBase
     private readonly ILogger<RestfulRegistryRepositoryController> _logger;
     private readonly HttpClient _httpClient;
     private readonly RestfulRegistryRepositoryService _restfulRegistryService;
+    private readonly IVariantFeatureManager _featureManager;
 
-    public RestfulRegistryRepositoryController(ILogger<RestfulRegistryRepositoryController> logger, HttpClient httpClient, RestfulRegistryRepositoryService registryRestfulService)
+    public RestfulRegistryRepositoryController(ILogger<RestfulRegistryRepositoryController> logger, HttpClient httpClient, RestfulRegistryRepositoryService registryRestfulService, IVariantFeatureManager featureManager)
     {
         _logger = logger;
         _httpClient = httpClient;
         _restfulRegistryService = registryRestfulService;
+        _featureManager = featureManager;
     }
 
     [HttpGet("document-list")]
     public async Task<IActionResult> GetDocumentList([FromQuery] string? id, [FromQuery] string? status)
     {
+        if (!await _featureManager.IsEnabledAsync("RestfulRegistryRepository_Read")) return NotFound();
+
+        var requestTimer = Stopwatch.StartNew();
+        
         var entries = _restfulRegistryService.GetDocumentListForPatient(id, status);
+
+        requestTimer.Stop();
+        
+        _logger.LogInformation($"Completed action: document-list");
 
         if (entries.Success)
         {
+            _logger.LogInformation($"Successfully retrieved {entries.DocumentListEntries.Count} entries in {requestTimer.ElapsedMilliseconds} ms");
             return Ok(entries);
         }
+
+        for (int i = 0; i < entries.Errors.Count; i++) _logger.LogError($"\n######## Error #{i} ########\n ErrorCode: {entries.Errors[i].Code}\n Message: {entries.Errors[i].Message}");
 
         return BadRequest(entries);
     }
@@ -40,26 +57,52 @@ public class RestfulRegistryRepositoryController : ControllerBase
     [HttpGet("document")]
     public async Task<IActionResult> GetDocument([FromQuery] string? home, [FromQuery] string? repository, [FromQuery] string? document)
     {
+        if (!await _featureManager.IsEnabledAsync("RestfulRegistryRepository_Read")) return NotFound();
+
+        var requestTimer = Stopwatch.StartNew();
+
         var entries = _restfulRegistryService.GetDocument(home, repository, document);
+
+        requestTimer.Stop();
+
+        _logger.LogInformation($"Completed action: get-document");
 
         if (entries.Success)
         {
+            _logger.LogInformation($"Successfully retrieved document with id: {entries.Document.DocumentId} in {requestTimer.ElapsedMilliseconds} ms");
             return Ok(entries);
         }
+
+        for (int i = 0; i < entries.Errors.Count; i++) _logger.LogError($"\n######## Error #{i} ########\n ErrorCode: {entries.Errors[i].Code}\n Message: {entries.Errors[i].Message}");
 
         return BadRequest(entries);
     }
 
     [HttpPost("upload")]
-    public async Task<IActionResult> UploadDocument([FromBody] DocumentReferenceDto value)
+    public async Task<IActionResult> UploadDocument([FromBody] DocumentReferenceDto documentReferenceDto)
     {
-        _restfulRegistryService.UploadDocumentAndMetadata(value);
-        return Ok("ok");
+        if (!await _featureManager.IsEnabledAsync("RestfulRegistryRepository_Create")) return NotFound();
+
+        var requestTimer = Stopwatch.StartNew();
+
+        var uploadResponse = _restfulRegistryService.UploadDocumentAndMetadata(documentReferenceDto);
+
+        requestTimer.Stop();
+
+        if (uploadResponse.Success)
+        {
+            _logger.LogInformation($"Successfully uploaded document and metadata in {requestTimer.ElapsedMilliseconds} ms");
+            return Ok(uploadResponse);
+        }
+
+        return BadRequest(uploadResponse);
     }
 
     [HttpPut("update")]
     public async Task<IActionResult> UpdateDocument(string id, [FromBody] DocumentReferenceDto value)
     {
+        if (!await _featureManager.IsEnabledAsync("RestfulRegistryRepository_Update")) return NotFound();
+
         _restfulRegistryService.UpdateDocumentMetadata(value);
 
         return Ok("ok");
@@ -68,6 +111,8 @@ public class RestfulRegistryRepositoryController : ControllerBase
     [HttpDelete("delete")]
     public async Task<IActionResult> DeleteDocument(string id)
     {
+        if (!await _featureManager.IsEnabledAsync("RestfulRegistryRepository_Delete")) return NotFound();
+
         _restfulRegistryService.DeleteDocumentAndMetadata(id);
         return Ok("ok");
     }
